More changes to the inner interpreter and CACTH to behave better in some artificially constructed abusive use cases, such as: variable cc : xx cc @ .s catch . ; ' xx cc ! xx
In 'real' threaded code if there is a check for underflow/overflow that can catch runaway cases like the one shown above.
But since we are running most of the system in C the underlying 'C' machine can also run out of stack space and crash, so I am trying to move stuff to the bounds-checked Forth stacks.
Also the previous version of the inner interpreter does not always seem to work with the CATCH implementation. Basically threaded code using that interpreter cannot return to CATCH (which is implemented in C in this system).
I left the code in for now (#ifdef-ed out) in case I am overlooking something obvious (it really is very late at night -- or rather early in the morning).